+++ examples/editor/src/ModalsMap.jsx
@@ -1,5 +1,7 @@
 import { ModalsMap as VideoModalsMap } from 'wix-rich-content-plugin-video';
+import { ModalsMap as SoundCloudModalsMap } from 'wix-rich-content-plugin-sound-cloud';
 
 export default {
   ...VideoModalsMap,
+  ...SoundCloudModalsMap,
 };