+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 
 ## [Unreleased]
 
+### :rocket: New Feature
+
+- `soundcloud`
+  - [#92](https://github.com/wix-incubator/rich-content/pull/92) `plugin-sound-cloud` implemented
+
 ### :bug: Bug Fix
 
 - `html`